[arm] Dramatic improvements to disassembly/decompiling!
It's now possible to peer into the target's memory and dump words (du),
disassemble code (dis), or decompile forth code (dec).

What is *really* cool is that dis and dec will switch back and forth.
Any time either one sees a target code field, it will print out what
kind of code field it is (if it recognizes it) and then start
disassembling or decompiling - whichever makes sense.

To see it in action, try this (you do *not* need a Pico for this
experiment):

  ./muforth -f target/ARM/board/raspberry-pi-pico.mu4
  @ram dis

and just keep pressing Return (or hold down the key). There are only a
couple of colon definitions in there, but if you are patient you'll see
them. ;-)

While dumping, disasming, or decomping, single key presses will also
change the mode:

  d => dumping
  e => decompiling  (would 'f' or ':' be better?)
  i => disasming    (i for instructions)

Enjoy!