Solved the decompile inconsistency for the stm32f0-discovery board and likely any others. Related to #31, and this is happening because muforth now includes "special" example code in the board loader file.
This is wrong.  It violates the principle of least astonishment for new
users, as they will be left scratching their heads trying to figure out why
flash has odd behavior.  The solution is simple; create new files with just
enough code to load and chat with a board that still has the factory demo
code on it.  Clearly label the loader files with the word "factory".

I've gone ahead and done that in this repo and a petition has been made to daf.

	modified:   ../beta-testing/stm32f0/dec-f051.txt
	new file:   README.md
	new file:   stm32f0-factory-discovery.mu4
	new file:   stm32f072b-factory-discovery.mu4
	new file:   stm32f3-factory-discovery.mu4
	modified:   stm32f4-discovery.mu4
	new file:   stm32f4-factory-discovery.mu4
	new file:   stm32f429i-factory-discovery.mu4
	new file:   stm32l-factory-discovery.mu4