[riscv] A few fixed to RISC-V code
I have to solve this "which colon compiler are we in" problem once and
for all.

Right now we are back to having to define assembler macros very
carefully. They have to be defined in the __host context, not __meta. So
you have to switch in and out to make this work.

This change has resulted from trying to simplify *another* place where
we have to be careful about which mode to return to: asm{ and }.

Maybe ; should do the same thing? The issue is that the ; that ends an
assembler macro returns us to the host forth (__host) rather than to the
interpret state we started in. Argh.

For now, the RISC-V chat code compiles again. But this problem really
needs a clean solution. It's one of the meta-compiler bugbears that I
have wrestled with for years!