[riscv] Fixed breakage introduced when I changed interact code
A while back I changed the ARM interact code and decided to make the
same changes to the RISC-V code, but I forgot to change the chat and
openocd implementations to match.

This commit fixes that oversight.

NOTE: ARM and RISC-V interact should really be the same code. I'd like
to make the two targets as similar as possible. (future work ;-)