Remove bad docs (faux pdf's that were really html) Add CONTENTS file
Note that we have mixed interest in the Espressif chips and currently
have no supporting code for them beyond a few equates files.

We're collecting docs so we can take a look and see if we want to support
them at all; the one that looks the most interesting to some of us is
the new 8685 because of its RISC-V core.  Sadly these look like they'll
be have availability delayed until early 2023, and at this point the
only doc we've found is the datasheet. :/

	new file:   mu/target/ESP/doc/0a-esp8266ex_datasheet_en.pdf
	new file:   mu/target/ESP/doc/0a-esp8285_datasheet_en.pdf
	deleted:    mu/target/ESP/doc/0a-exp8266ex_datasheet_en.pdf
	deleted:    mu/target/ESP/doc/4a-esp8266_at_instruction_set_en_v1.5.4_0.pdf
	new file:   mu/target/ESP/doc/CONTENTS
	new file:   mu/target/ESP/doc/ESP8266_Pin_List_0.xls
	new file:   mu/target/ESP/doc/esp32_bluetooth_architecture_en.pdf
	new file:   mu/target/ESP/doc/esp32_bluetooth_networking_user_guide_en.pdf
	new file:   mu/target/ESP/doc/esp32_datasheet_en.pdf
	modified:   mu/target/ESP/doc/esp32_technical_reference_manual_en.pdf
	modified:   mu/target/ESP/doc/esp8266-technical_reference_en.pdf
	new file:   mu/target/ESP/doc/esp8685_datasheet_en.pdf