Add kinetis doc/, link in README_Included_Documentation.md
NB: NXP is on the back burner for us atmo. At the time our primary focus
was the KL25Z, so the included documentation and most of the work in the
Kinetis repo reflects this.  There are mixed feelings regarding NXP, and
their cheap dev boards haven't seemed to gain much traction with the
hobbyist community.  But, if a Kinetis chip is your huckleberry, there's
at least a blueprint to follow here.

	modified:   README_Included_Documentation.md
	new file:   mu/target/ARM/kinetis/doc/CONTENTS
	new file:   mu/target/ARM/kinetis/doc/FRDM-KL25Z-Pinouts-Rev1.pdf
	new file:   mu/target/ARM/kinetis/doc/KL25P80M48SF0.pdf
	new file:   mu/target/ARM/kinetis/doc/KL25P80M48SF0RM.pdf