[arm] Testing QSPI flash (W25Q16JV) on Raspberry Pi Pico
The Pico SDK contains "stage2 boot" code that uses undocumented Winbond
W25Q16JV behavior (regarding the "mode bits" in the Fast Read Quad I/O
command (EBh)).

After reading the datasheet for that chip, and the datasheets for about
five other chips (Micron, ISSI, Cypress, GigaByte...) I was rather
confused, and I wanted to understand exactly what was going on.

I decided to try *bitbanging* the QSPI interface (by connecting it to
the RP2040 SIO block) and discovered that, yes, using "Ax" mode bits
works, but it's not documented *anywhere* in the Windbond universe. I
would prefer to run this chip in a totally documented way instead - but
that's a different piece of code and a different story.

To try what I tried, do this (I have modified the Pico board load file
to load my QSPI test code):

  ./muforth -f target/ARM/board/raspberry-pi-pico.mu4
  pico
  init-spi
  1000 read16
  sp-reset
  1000 qread16
  sp-reset
  2000 xip-start drop
  1000 xread16
  sp-reset
  2000 xip-end drop
  1000 read16
  sp-reset

read16 reads 4 words from a flash address, using the 03h read command,
and single-wire SPI mode.

qread16 reads 4 words from a flash address, but using the EBh QSPI read
command instead. It sends Fx to the chip as the mode bits (as documented
in the Winbond datasheet). This does *not* put the chip into "continuous
read" mode.

xip-start reads a single word from a flash address, and tries to put the
chip into continuous read mode, using Ax mode bits.

xip-end reads a single word from a flash address, but sends Fx mode bits
to try to return the chip to "normal" SPI command mode.

xread16 reads 4 words from a flash address, but assumes that the chip is
in "continuous read" mode, and sends Ax mode bits to keep it in
continuous read mode.

sp-reset clears the stack to make it easier to compare the results of
the reads (they are reading from the same address - 1000 - and should
return the same data).

Hope this helps to illuminate this murky corner of the RP2040 and the
Pico!