Change the name again. b0g25 maps more sensibly to the register name.
     : init-b0g25 5 IO_BANK0_GPIO25_CTRL ! ;

The register names of the Pico are just a ridiculous hot mess as it
is ... but as always we're keeping them (the lua equates rendered from RPi's
svd file) because we want as close a mapping as possible to the canonical
documentation. (In the past we've used a word, aka, to rename words/registers
etc when confronted by extremely silly or outright bad vendor naming choices.)

Segue: I suppose if you're just using micropython to whack on the pico and
blink some lights, the documentation seems fine; after all, it's as much of a
manual for their SDK as it is a datasheet.

But:

     from machine import Pin
     led = Pin(25, Pin.OUT)

     led.toggle()

doesn't teach you much of anything except how to type the above.  It certainly doesn't
teach you a damn thing about what's going on under the hood, and if you're working
with hardware the last thing you need are super high-level hand-wavey abstractions.

	modified:   pico-testing.mu4