[riscv] Working serial chat for Longan Nano (GD32VF103) RISC-V board!
This is a big milestone for muforth!

(It probably also breaks the HiFive1 support, so I have to look at that
next.)

I have only tested the serial bootloader. If you have a USB-C cable you
can try the DFU.

The easiest way to get started with this code is the following:

If using the serial bootloader, connect an external USB-serial adapter
to the power, ground, TX, ard RX pins on the Nano. Plug in the adapter.

If trying the DFU method, connect the Nano with a USB-C cable.

Press the reset and boot buttons, release the reset button, then release
the boot button. (In the commands below, substitute "dfu" for "boot" if
you are going that route.)

  ./muforth -d bootloader -f target/RISC-V/load-chat-nano.mu4
  boot
  flash-image
  verify

Quit muforth (^C will work) and reload like this:

  ./muforth -f target/RISC-V/load-chat-nano.mu4
  chat
  verify

To test that it's really working, try disassembling the chat code:

  @flash dis

And to try running some code (not Forth code, just simple assembler):

  flash-leds rx

The RGB led should quickly run through its colors, just as it does after
reset.

If you want to compile more code to try, do something like this:

label my-code
   <assembler code here>   ret  ;c

Then:

  my-code rx

You can pass values on the stack to and from your word, just as if it
was a Forth code word.

Enjoy!!