[riscv] Compatibility hack for existing HiFive1 chat code
Until I get the new chat code working on the HiFive1, there is a "hack"
that can be used to continue to chat with the board, execute code, etc.

The hifive1 board file now defines "hifive-compat" which slightly
changes how chat-host talks the the board.

To try it out, maybe try this:

  ./muforth -f target/RISC-V/board/hifive1.mu4 \
      ram -f target/RISC-V/kernel-itc.mu4 -f target/RISC-V/hifive1/led.mu4
   chat
   led-init
   red set
   blue on
   green on