[riscv] Changed an SPI-related name so I could keep rx for asm exec
I like "rx" as a shorthand for "remote execute". I sometimes use this
for executing non-Forth assembler words ("label" definitions).

Because I had define "rx" and "tx" in the Hifive1 SPI code (for flashing
the external SPI Nor flash on the board) I temporarily renamed the
remote exec rx to something else.

But I like rx too much for this, so I renamed the SPI routines, and now
rx is available again to use to execute code.