[riscv] Made it easier to call code that defines reset behaviour
The interact word "reset" now calls the code whose address is contained
in the (host) variable 'reset.

I'm setting this, in the chat code, to the beginning of the code that is
executed as reset time.

This makes it easier to test new code (eg, chat code that is loaded into
ram) or to run a ram-resident version of the chat code to use while
writing a new version of chat to the flash.