[riscv] Moving toward common code for flashing targets
Common, at least, between ARM and RISC-V. ;-)

I added "start flash" and "end flash" commands to the set of
"interact" commands. This allows debug/chat code to do any mode
switching or locking/unlocking before and after the actualy erasing and
flashing process.

Making the code shared required moving some things around, and I decided
to give the RISC-V SPI flashing code (written in Forth) its own
mini-kernel, because with the normal Forth kernel it was embarrassingly
big.

I have more testing to do, but things on the ARM side are close to ready
to be committed as well.