[arm] (WIP) Mostly working basic USB firmware for STM32F0 and F3 devices
I claim that it's "mostly working" because it enumerates with almost
100% relibility when executed from RAM (using ST-LINK), and with good
reliability when flashed and executed at reset (by pushing the reset
button) but only when when *both* ST-LINK and user USB cables are
connected...

With only the user USB cable connected it neither successfully
enumerates at plug-in or when reset is pressed.

I think that the core firmware handling of control requests is working,
but somehow the timing of the disconnect/connect/reset/enumerate is
finicky.

More work is forthcoming; I just wanted to check this in as
documentation and as a PoC.