[arm] Fixed long-standing bug in (0branch) !!!
Thanks again to Eelco for catching this!

The bug: I was not correctly (or at *all*, really) checking the value of
top in (0branch). Instead I was moving it to w (why?) and then branching
based on the Z condition code.

Oddly enough, this mostly worked for the Forth code I was writing,
because I was generally, just before the if/while/until (which got
compiled as (0branch)), doing some kind of computation (eg, "and") or a
comparison that set the condition codes, and then (0branch) got lucky
and and re-used those settings!

I was hoping that this fix would cause my USB firmware to start working
reliably, but nothing changed.

Anyway, Forth code should be much more reliable now - esp if it computes
flag values and passes them along to other code which much later does an
if/while/until.