[arm] Simplified picoboot support
Seems like it's possible to read the flash using the picoboot read
command, but you have to be careful.

du will work, but only on addresses that are a multiple of 256.

However, since the flash verify command reads the flash in multiples of
256, it will work nicely with the current picoboot "settings".

It's my conjecture that picoboot reads the flash using SPI commands, not
using XIP, so turning XIP off is the right thing to do.

However, this means that it's not possible to either *execute* code in
the flash, or read it using target-resident Forth memory access words
(like @ and @+).

Once we have a proper stage2 for the flash this will get better. For now
it's necessary to continue our explorations using ram-resident code.