[arm] It's now easy to add a CRC32 to your Pico's stage2 loader
The word

  append-crc32

will calculate a Pico-bootrom-compatible CRC over the first 252 bytes of
the flash image (@flash to @flash 252 +) and store it in the last 4
bytes of the page. This makes it easy to compile code into a boot
"stage2" and add the necessary CRC to it so the Pico bootrom will jump
to it.