[arm] Experiments with XIP on the Raspberry Pi Pico
I was able to get both the basic SPI read instruction (03 hex) and the
"fast read quad output" instruction (6b hex) working, but I was unable
to get the "fast read quad i/o" instruction (eb hex) working.

The latter requires mode bits, which allow entry into, or exit from, a
"continuous read" mode. This is a bit esoteric and non-standard, and, in
fact, the Pico SDK stage2 uses this continuous read mode, even though it
is *not* specified as supported in the Winbond W25Q16JV datasheet!

I wanted, if possible, to eschew depending on undocumented behavior, so
I limited myself to trying to get the XIP peripheral on the RP2040 to
execute "eb" reads without going into continuous read mode. Since the
XIP peripheral has a curious - and rather hacky! - design, I was unable
to get it to work.

I'm tempted to stop here and create my own stage2 using the "6b" read
instruction.