[core] Improved and refactored string compilation primitives
I hope things are clearer and more readable now.

I also added a second compiled string literal primitive, (z") . Most of
the time we are only interested in the string address, not its count, so
after a string literal there was a hidden "drop" compiled into the
dictionary to drop the length.

With two primitives it's possible to save a tiny bit of space in the
dictionary, since  z"  (which compiles (z") ) is much more common than
" .