[core] WIP - mixed 32/64 implementation
I'm checking this in because there are some interesting ideas in here
for dealing with, eg, a 32-bit wide heap even on 64-bit machines, using
relative addressing. There are interesting questions about when you want
to deal with relative vs absolute addresses, when to convert them, etc.
I came to the conclusion that it made the most sense to always push
absolute addresses onto the stack - and @ and ! and friends would expect
absolute addresses - only converting them to relative form when writing
them into or reading them from a heap cell.

NEXT would be a bit slower because it has to do this conversion - when
reading an xt (a 32-bit relative pointer), it has to make it absolute (by
adding the heap origin address) before fetching the code field - which
is also a relative address, but this time relative to mu_do_colon;
chosen carefully so that colon words would always have a code field with
all zeros, so it would be easily recognizable. ;-)

I've decided to pursue a different approach. Instead of having a nice
compact heap on all platforms, instead let's favor the address size of
the machine. On a 32-bit machine, heap cells - which mostly contain
addresses - are 32 bits wide. And on a 64-bit machine they are 64 bits.
We can write a very natural NEXT that uses machine pointers in most
obvious and efficient way.

The tricky part comes with dealing with *values*. I want to keep the
64-bit computational nature of muforth intact, so the stacks will remain
64 bits wide on all platforms. This works cleanly on a 64-bit machine:
the stacks and the heap are both 64-bits wide. On a 32-bit machine,
we have to store values in two consecutive cells, and be careful about
alignment or do 64-bit fetches and stores in two 32-bit pieces. I'm not
too concerned about this because 32-bit machines are not the "intended"
host platform for muforth; I'm much more interested in a stellar
experience on 64-bit machines. But if I can get it to work reasonably
well on 32-bit machines, great.

Some of the changes I've made while experimenting with the
relative-addressed heap still apply, and there are also some changes in
the way that values are communicated between C and Forth that I want to
keep, so I'm going to start from here as I continue on this adventure
into the heap!