[core] WIP, continued. I got the 64-bit version working again!
But that's the easy one, because values and cells are the same size.

The tricky platform is a 32-bit machine.

I'm also confused about what to call the words that read, write, and
compile cell-sized values. I started with cell, cell@ cell! but changed
my mind and started to call them p, p@ p! because they are only used for
*pointer-sized* values. Cells hold pointers; vals hold values. Cells are
32 or 64 bits; vals are always 64.

I'm going to try compiling the code on a 64 bit machine with -m32 and
see if I can get it to work.

Wish me luck!