[core] WIP - snapshotting what I have before a name change
Henceforth I want to use the term "cell" to mean a user-visible storage
slot. A variable is one cell; an array is made of cells. A cell is
always 64 bits wide.

On 32-bit machines the pointers - code, link, and execution token (xt)
pointers - are 32 bits, unlike cells. For this I am going to have a
separate type and a handful of separate operators. This type, which
represents an implementation detail and not something "normal" users
will be interested in, is going to be called an "addr".

On 64-bit machines, cells and addrs are the same. This only affects the
structure of the heap on 32-bit machines.