[core] compile, is the new name for compiling a cfa into the dict
It used be just  ,  but that copies a cell-sized value into the
dictionary.

To compile a cfa (or execution token - xt) into the dictionary we need
to copy an addr-sized value into the dict.  addr,  does this, but having
a synonym that says *what* we are doing, not *how* (ie,  compile,  )
is useful.