Made the beginning of startup.mu4 much more readable!
The very first thing that happens is that I define the comment
character, ( .

Using that I can walk the reader through the initial machinations, which
are otherwise quite an inscrutable puzzle!

Even with the comments, following along isn't trivial. But at least now
you have my voice in your head to guide you. ;-)