Make \ more efficient and maybe more readable!
\ is a weird word. It is like ANS's postpone, and does one of two
things, after parsing the following token:

* search .compiler.; if token is found there, compile it into the word
  we are currently compiling (it is most likely another compiling word,
  and we want to include in its behavior the behavior of *another*
  compiling word)

* if not in .compiler, search .runtime. (this really searches both
  .runtime. and .forth. , because .runtime. is chained to .forth. ); if
  not found, complain; otherwise, compile code into the word that we are
  currently compiling (probably a compiling word!), to *later* compile the
  word we just into yet *another* word. ;-)

Yes, it's complicated.

And the previous implementation of the second clause (the .runtime.
part) was both hard to read and inefficient in what it compiled.

Now, if the word is not in the .compiler. chain but is found in the
.runtime. or .forth. chains, the following code is compiled:

  compile <runtime-or-forth-word>

It is compiled by the code in \ , which now reads as follows:

  compile compile  compile,

I hope that makes sense. It's a brain teaser, for sure. As is any aspect
of Forth that straddles several time phases (compiling, compiling of
compiling words, execution).