[riscv] Accidentally created an impossible pseudo-instruction
After defining

  sltz  ( set if register less than zero)

which compares a register to x0 (the zero register), I defined a version
with the compared registers swapped; I *should have* called this:

  sgtz  ( set if zero less than register)

but instead I called it "sgez" - set if greater than or equal to zero!
Which is totally wrong.

So sgez is gone, replaced by the *correct* pseudo-instruction sgtz!