[core] Finally solved the show/hide conundrum!
There is a tradition in Forth systems that colon definitions are somehow
hidden until they are complete. The idea is that you can redefine a
colon word, using the same name, and in the text of the redefinition the
name will refer to the *old* version; but as soon as the ; is reached,
the new definition becomes the one that  find  will return.

In fig-FORTH this was done by having a "hidden" bit in the name field.
muforth has had a hidden bit as well: I added and removed it several
times!

In muforth, names need to be hidden for only two reasons:

 * the name is a fake name ("muchain") that is used in the vocabulary
   chain linking implementation;

 * the name has been undefined by the user, using  undef

In both of these cases, I use a length byte of 0 instead of a hidden
bit.

So, what's a simple and elegant way to hide a colon definition while it
is being defined? Martin Tracy's zenFORTH suggested one: delay linking
the name into the dictionary chain until it is complete.

Until this commit, I used this technique in muforth, but in a clumsy
way: the dictionary code that creates new name - which is written in C -
*would* immediately link the new name into the chain; then, in Forth, I
would unlink it, but remember where it used to be linked (this word was
called, unsurprisingly, "hide"); and then when ; executed and the colon
word was complete, "show" would re-link the name, using its saved
information.

This has bugged me for a long time, but I wanted to avoid adding more
variables that need to be shared between C and Forth (because that
necessitates the use of addr, which is potentially confusing), and I
hadn't realized until today that a simple solution was staring me in the
face.

Now, when the C code creates a new name, it links the name to the chain,
but doesn't link the chain to the name. It remembers, in two static C
variables, the name and the chain. Later, the word  show  which is
implemented in C, simply finishes the job and points the chain to the
new name. It's simple and elegant.

Now I can move on to other muforth warts that have been bothering me for
ever, like the metacompiler mess! ;-)