[msp430] Made flashing of boot code and vectors more flexible
The F5529 boot/chat code exceeded the space in the last 512 bytes of the
flash. There was a lot irritating code to get the clocks and such
configured.

I changed the location of the boot code - @boot - to be device-specific.
I was also careful to re-read the datasheets and make sure that @vectors
and #vectors were accurate.

Lastly, the flash code was changed to be smart about the fact that the
vector page (always the last page in flash) and the beginning of the
boot code may no longer coincide, so it has to be careful about erasing
the vector page separately if necessary.