[arm] Removed @ ! and friends from meta vocab (interact code)
I had added these to make interacting with a bare target less painful.
It's as if you had a Forth kernel running on the target, and could do
memory reads and writes.

The problem is that it interacted badly with the meta-compiler code.

I think it's better to simply *get* a working kernel onto the target and
use that to interact. (This is already available on the ARM target.)

This is the first of a few changes - one a bit more radical - that
should make the meta-compilers more robust and easier to think about.