[arm] Fixed a Very Bad Bug in the target "variable" definer
variable was broken such that it worked when compiling code into RAM,
but totally failed when compiling code into flash!

This was the main reason that my USB code for the STM32 didn't work! I
was able to get it to enumerate somewhat reliably when testing
ram-resident code, but once I flashed it, it just sat there.

This fixes that, and my sincerest apologies to anyone who has been
struggling because of this!