[arm] O frabjous day! Working USB firmware for STM32F303/F072!!!!!
Late last night I realized that I the "variable" defining word was
wrong, and wouldn't work for code compiled into flash.

I fixed it, re-compiled the USB test code, and it immediately worked!

This is after trying over and over to get it to work, and basically
giving up on ST and the STM32. It turned out to be a stupid simple
mistake on my part.

I've made a few small changes to the USB code, including commenting out
the possibly offensive "manufacturer string" and replacing it with
something banal.

I've also made the USB code the "out of the box" code that the STM32F072
Discovery and F3 Discovery board files load.

You too can try this at home!

  ./muforth -f target/ARM/board/stm32f072b-discovery.mu4

OR

  ./muforth -f target/ARM/board/stm32f3-discovery.mu4

then

  jtag
  flash-image
  verify

Reset the board, plug into the "user USB" jack, and watch your favorite
OS enumerate the board. I've had success with Linux and Windows 10.

Pressing the reset switch will disconnect from the USB, and the board
will re-enumerate.

It doesn't do anything more than that, but the hard part was getting to
this point. Adding code to process vendor requests to turn it into a
chat target is pretty easy.

Super exciting!!!!