[msp430] Big improvements to MSP430 disassembler and a new decompiler
Thanks to work done on the ARM target that was easy to port over, thanks
to the uniformity of ITC Forths, the MSP430 "dis" command will now
seamlessly switch between disassembly and decompiling of colon
definitions.

In the disassembly it also prints equates by *name* rather than by
value, so it is much easier to see what the code is doing! Try it on the
chat support for one of the chips to see what I mean? (Those files are
in mu/target/MSP430/device/<something>/chat-support.mu4)