[arm] Added experimental code to picoboot for executing assembler
Not sure if I want to do this; it's a bit of a proof of concept.

The idea was to be able to read memory by executing code, rather than
using the PICOBOOT commands, which are very limited: they won't read ROM
beyond 8KiB; they won't read i/o space at all; and they won't read the
flash except on 256-byte boundaries.