Fixed untoken's tendency to screw up line numbers
When using untoken to back up over an already-parsed token, we have to
be careful. If the token was followed immediately by a newline, the line
number will have been incremented. If we don't reset it, our line
numbers will get out of whack, and error messages will refer to the
wrong source line (which is how I discovered this bug).