Removing an odd relic - target/common/memory.mu4 - which was never used
It has some weird and interesting ideas in it, and usages that suggest
that it was derived from the V25 compiler - \o in particular.

I'm not sure what to do with this, so I'm deleting it from the tree but
leaving behind this message so it can possibly be found if someone is
interested in it.