[8051] WIP -- The beginnings of target support for the 8051!
The assembler is incomplete; type

  assembler words

to see which words are defined. You can easily play around by typing

  __asm

to start the assembler; any compiled instructions are simply printed out
(they are not compiled into memory, though this is a trivial change,
since all the memory image support has been baked in).

Target du does *not* work; executing du with a target address will
crash.

Mostly this is a proof-of-concept and a testbed for the assembler.