[8051] Assembler is complete!
But also completely (or at least *mostly*) untested!

It also lacks the control structure compiling words, but those are easy.

This is an architecture whose instruction encoding seriously has more
exceptions than rules! Factoring the code to implement an assembler was
a tedious and iterative process, but I think the result isn't too bad.