[8051] Removed call and jump "shorthands"
I have used c as a shorthand for call, and j as a shorthand for jump in
a bunch of the targets.

The 8051 assembler, unfortunately, needs "c" for carry.

So I just got rid of the shorthands entirely. You have to use acall,
ajmp, sjmp, lcall, or ljmp, as necessary, in your 8051 code!