[8051] Chat improvements
Now the chat code running on the chip uses bank3 registers, and the host
puts the flash command and keys directly into r3 to r5 by writing to
their ram aliases.

I've decided that the simplest way to get values to and from pieces of
code that get executed is to have the host read and write the ram
aliases of the registers. For this reason I changed the chat code to use
bank3 registers, leaving all the others free for user code.