[8051] Compile chat code as EFM8 bootloader
Right now flash-image will send a lot of FFs to the target, but the chat
code ignores these and doesn't program them.

The generic ("core") flash code needs to be modified so that it can do
special target-specific things when flashing and verifying, to, eg,
flash or verify the interrupt vectors, which are sometimes separated by
unprogrammed space from the code being flashed.