[8051] Disassembler, baby!
For some reason it's really hard to make sense of the way the opcode
space is arranged. Writing *both* the assembler and disassembler for
this device were both rather challenging - surprising given the seeming
simplicity of the instruction set.

But the encoding has more exceptions than rules! Maybe there is a more
elegant way to do it and I just missed seeing it.