[arm] Modified ST-LINK support to be independent
I want to be able to talk to running code - even code running in RAM -
on a Discovery board while I am loading, flashing, or interacting with a
completely different target - AVR or 8051, say.

It's now possible to load some code into the STM32's ram, start it, then
reload muforth with *just* the ST-LINK support, re-connect, and interact
(by changing values in RAM) with the running code.