[avr] Reworked build to be more flexible, esp about SPI programming
It used to be that the only way to program an AVR using muforth was by
using as JS16 USB board that only I own!

I wanted to fix this, so I have worked on STM32 code running on two
Discovery boards (F072 and F303) to use as programming "pods", to
program Atmel 8051 and AVR parts via SPI.

You can now define "via-s08" to get the old S08 code, or "via-disco" to
get the STM32/Discovery code.