[avr] Factored chat code to share as much as possible
I want to be able to "test" the core ideas of the AVR chat code without
worrying about the details of I2C working.

I've factored the code for both serial and I2C chat to share as much as
possible. This way I can (hopefully) test that starting and stopping an
application, and erasing and reflashing the app code, works over a
serial connection. UARTs are easier to deal with.

Once everything is working, hopefully the details of the I2C won't be
too overwhelming and everything will "just work"!