[arm] Fix Pico board load file that it actually loads!
It was trying to load the stage2 flash code, which had a hack in it to
call next, but next wasn't defined because there isn't a kernel
loaded...

I just commented out the loading of stage2. We don't need it yet, and
can't use it yet. Until a few more pieces are in place...