[arm] Picoboot improvements!
The biggest one: PICOBOOT limits reads; it's impossible to read the 2nd
half of the ROM, or any of the i/o space.

But PICOBOOT allows us to execute code. So why not write a tiny routine
that reads an arbitrary bit of the memory space, copy that into ram, and
then use PICOBOOT to read the ram?

That's exactly what we do now. ;-)