[arm] Almost everything to do with flash is BROKEN on the Pico!!!
This commit works around all of the problems that I have encountered.
Here is a list:

PICOBOOT
========

* NONE of the flash-related commands work, at all. Period. This includes
  enter-xip, exit-xip, erase, write (to a flash address).

* Reads from flash only work if the starting address and length are both
  multiples of 256. Otherwise, boom.

* Reads from the 16 KiB ROM only work if they are from the bottom 8 KiB.
  Otherwise, boom.

* Reads from i/o space ... boom.

Workarounds: At connection time, two routines are copied over to the
Pico's ram: a generic memory read routine without any of the above
limitations, and a stage2 routine that sets up the Pico's XIP peripheral
so that flash can be read like any other memory.

XIP
===

The XIP peripheral is very badly designed and, frankly, broken. The only
way to make use of the most efficient QSPI read instruction - 0xEB, or
"fast read quad i/o" - is to use a hacked-on hardware mode that makes
use of "continuous read" mode bits that, while they HAPPEN TO WORK on
the Pico's SPI flash chip, are undocumented.

Trying to use the EB command without the mode bits - which the flash
chip is totally capable of (I proved it by bitbanging the QSPI via the
SIO) - is impossible to do with the Pico's XIP. It just doesn't seem to
do the right thing with the dummy cycles. I tried and tried and finally
gave up.

Workarounds: I'm currently using the "fast read quad output" - 0x6B -
command, which isn't nearly as efficient, but I didn't want to use an
undocumented mode of the flash chip, and I likewise didn't want to trust
that the broken XIP would actually fucking *work*.