[core] Colorized! And any printed text goes by default to stdout
This is an experiment, and the colors are just a grab-bag. I'm not crazy
about them, and you might not be either.

If the whole thing annoys you, edit mu/startup.mu4, find the definition
of "warm" and change +color to -color.

If you have some fave colors that you would like to use, find the
definition of info-text in mu/startup.mu4 and put in the colors you
prefer.

If you want to use muforth to generate text, in the past you had to fool
around with code like "out-channel preserve >stdout"; all that is now
unnecessary. By default, *all* text printed by *any* code goes to
stdout; only colorized "special" text - info, status, delimiter,
warning, error - goes to stderr.

As always, enjoy!