[avr] Improvements to IIC chat code (core)
Among other things, I realized that any read requires an I2C
"turnaround" from write to read; this means that the code has to wait
for a restart; and the way that code works can sometimes reset the
stack.

I was saving intermediate values onto the stack when reading 16-bit
values; this could fail. So this was changed to save the intermediate
values in registers instead.