[avr] Renamed rjmp to _rjmp and ljmp to _ljmp
These are the Forth helpers, not the assembler words.

The problem: if you were accidentally *not* in the assembler mode and
you executed either rjmp or ljmp, instead of an error it would compile a
weird new create/does> word (and consume a token from the input stream),
causing all kinds of unintended consequences!

Now it causes an error.