[avr] Added code in test/ to test I2C by using serial chat
Working so far. But the big test - whether it's possible to write to the
flash via I2C - is not possible because I'm running the I2C code as an
"app", and the app can't be running when re-flashing because that's
precisely the part of the flash that gets re-programmed! Silly me.

So I have to do an even *weirder* experiment: running serial (uart) and
i2c chat side-by-side!

This might be a bit of a challenge.