[avr, arm] Time to celebrate! Working flash programming over I2C!
I have tried off and on for several *years* to get this to work. I'm not
sure *exactly* what fixed the problem, but I have a hunch: I was
blinding popping the stack by a fixed amount when a certain "exceptional
event" happened, and I think the stack wasn't always balanced.

Now I have a very simple catch/throw: catch captures the SP and PC at a
certain point in the program (the beginning of the "chat" conversation),
and when something "bad" or unexpected happens, the program resets the
SP and jumps to that PC.

I tested the code in a rather unusual way: I ran the two chat/debug
stubs - one talking to the UART, and one talking to the I2C - in
parallel, using a trivial (and elegant!) multi-tasker. I had confidence
in the uart chat code, as I had proved that I could program the chip's
flash over that interface; so it was a trustworthy "agent on the inside"
that could tell me what was happening with the I2C.

The other thing I did was to capture and log the I2C status bytes. Every
time something happens - the chip see its slave address and a write bit,
or a byte was transmitted and acked by the master - the status register
returns a code describing what happened. So your code has to be a sort
of state machine (although I didn't write it that way) - which is where
the "exceptions" come in. If we are expecting one code, but get
something else, we bail out and start over by "throwing" to the catch.

It's quite a remarkable experience to be able to peer inside a chip via
two debug interfaces simultaneously.

Multi-tasking, even in the simple form found here, is remarkably
powerful.